+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # ChangeLog
 
 * add ja_jp translation (thanks to @A8T1N)
-* add Nature's Aura - aura
-* add Pneumatic Craft  - pressure / air
+* add Nature's Aura - aura (not test yet)
+* add Pneumatic Craft  - pressure / air (not test yet)
+* add Embers - ember (not test yet)
 * fix event node graph doesn't work in some cases
 * fix kjs recipe functions